Сделал модуляцию кода
